% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{add_letters}
\alias{add_letters}
\title{Add Tukey letters to existing ggplot object}
\usage{
add_letters(p, x, y, group)
}
\arguments{
\item{p}{A \code{ggplot} object}

\item{x}{variable to plot on x axis}

\item{y}{variable to plot on y axis}

\item{group}{grouping variable (to allow faceting)}
}
\description{
Add Tukey letters to existing ggplot object
}
\author{
Ethan Bass
}
