% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.color}
\alias{is.color}
\title{Check whether color specifications exists.}
\usage{
is.color(x, return.colors = FALSE)
}
\arguments{
\item{x}{Vector of any of the three kinds of R color specifications,
i.e., either a color name (as listed by
\code{\link[grDevices]{palette}colors()}), a hexadecimal string of the form
'#rrggbb' or '#rrggbbaa' (see rgb), or a positive integer i meaning
\code{\link[grDevices]{palette}()[i]}.}

\item{return.colors}{Logical: logical values (FALSE, default) or
returning colors (TRUE)}
}
\value{
Logical value (or colors)
}
\description{
Function to check whether all specified colors are
actual colors.
}
\note{
Adapted from plotfunctions packaage https://cran.r-project.org/web/packages/plotfunctions/index.html
}
\examples{
# correct color definitions:
is.color(c('#FF0000FF', '#00FF00FF', '#0000FFFF'))
is.color(c('red', 'steelblue', 'green3'))
is.color(c(1,7,28))
# mixtures are possible too:
is.color(c('#FF0000FF', 'red', 1, '#FF0000', rgb(.1,0,0)))

# return colors:
# note that 28 is converted to 4...
is.color(c(1,7,28), return.colors=TRUE)
is.color(c('#FF0000CC', 'red', 1, '#FF0000'), return.colors=TRUE)

# 4 incorrect colors, 1 correct:
test <- c('#FH0000', 3, '#FF00991', 'lavendel', '#AABBCCFFF')
is.color(test)
is.color(test, return.colors=TRUE)
}
\author{
Jacolien van Rij
}
