% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{add_letters}
\alias{add_letters}
\title{Add Tukey letters to existing ggplot object}
\usage{
add_letters(
  p,
  x,
  y,
  group = NULL,
  test = "tukey",
  type = c("global", "local"),
  where = c("box", "whisker")
)
}
\arguments{
\item{p}{A \code{ggplot} object}

\item{x}{variable to plot on x axis}

\item{y}{variable to plot on y axis}

\item{group}{grouping variable (to allow faceting)}

\item{test}{Which test to run for pairwise comparisons. Default is \code{tukey}.}

\item{type}{Whether to run separate tests for each facet (\code{local}) or
one (\code{global}) test with an interaction term between \code{x} and
\code{group}. Defaults to \code{global}.}

\item{where}{Where to put the letters. Either above the box (\code{box}) or
above the upper whisker (\code{whisker}).}
}
\description{
Add Tukey letters to existing ggplot object
}
\author{
Ethan Bass
}
