% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_letters.R
\name{boxplot_letters}
\alias{boxplot_letters}
\title{Create ggplot boxplot with compact letter display}
\usage{
boxplot_letters(
  data,
  x,
  y,
  fill,
  group,
  test = "tukey",
  type = c("global", "local"),
  where = c("box", "whisker"),
  raw = c("none", "points", "dots", "jitter"),
  pt_col = "slategray",
  ...,
  hjust = 0,
  vjust = 0,
  lab_size = 4,
  na.rm = TRUE,
  threshold = 0.05
)
}
\arguments{
\item{data}{A data.frame in "long" format.}

\item{x}{variable to plot on x axis.}

\item{y}{variable to plot on y axis.}

\item{fill}{column or color to fill boxplots}

\item{group}{A grouping variable (to allow faceting).}

\item{test}{Which test to run for pairwise comparisons. Default is \code{tukey}.}

\item{type}{If a grouping variable is provided, determines whether to run
separate tests for each facet (\code{local}) or one (\code{global}) test with
an interaction term between \code{x} and \code{group}. Defaults to \code{global}.}

\item{where}{Where to put the letters. Either above the box (\code{box}) or
above the upper whisker (\code{whisker}).}

\item{raw}{Whether to plot raw data and (if so), how. The current options are
\code{none}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{geom_dotplot}}, or
\code{\link[ggplot2]{geom_jitter}}.}

\item{pt_col}{Color of points, if raw data is plotted.}

\item{...}{Additional arguments to \code{\link[ggplot2]{geom_point}},
\code{\link[ggplot2]{geom_dotplot}}, or \code{\link[ggplot2]{geom_jitter}},
according to the value of \code{raw}.}

\item{hjust}{Horizontal adjustment of label. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{vjust}{Vertical adjustment of label. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{lab_size}{Label size. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{na.rm}{Logical. Whether to remove observations with NAs for the provided
factors (i.e. \code{x} and \code{group}) before plotting.}

\item{threshold}{Statistical threshold for significance. Defaults to 0.05.}
}
\value{
Returns the specified plot as a \code{ggplot} object.
}
\description{
Does pairwise comparison using \code{\link[stats]{TukeyHSD}} and produces
boxplots with compact letter display showing significance pairwise differences.
Letters are produced by \code{\link[multcompView]{multcompLetters}}. Plots are
produced by \code{\link[ggplot2]{ggplot2}}.
}
\details{
Allows group variable for faceting
}
\note{
Adapted from https://www.mathiasecology.com/code/add-tukeys-significant-letters-to-ggplots
}
\examples{
set.seed(1)
data <- data.frame("Category" = c(rep("Low", 10), rep("Medium", 10), rep("High", 10)),
                   "Value" = c(rnorm(10, 5), rnorm(10, 5.5), rnorm(10, 10)),
                   "Size" = c("Big","Small"))
boxplot_letters(data, Category, Value)
boxplot_letters(data, x=Category, y=Value, group=Size)
}
\author{
Ethan Bass
}
