% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_letters}
\alias{boxplot_letters}
\title{Create ggplot boxplot with Tukey HSD letters}
\usage{
boxplot_letters(
  data,
  x,
  y,
  fill,
  group,
  raw = c("none", "points", "dots", "jitter"),
  pt_col = "slategray",
  ...
)
}
\arguments{
\item{data}{A data.frame in long format}

\item{x}{variable to plot on x axis}

\item{y}{variable to plot on y axis}

\item{fill}{column or color to fill boxplots}

\item{group}{grouping variable (to allow faceting)}

\item{raw}{whether to plot raw data and (if so) which format}

\item{pt_col}{Color of points, if raw data is plotted.}

\item{...}{Additional arguments to \code{\link[ggplot2]{geom_point}},
\code{\link[ggplot2]{geom_dotplot}}, or \code{\link[ggplot2]{geom_jitter}},
according to the value of \code{raw}.}
}
\description{
Allows group variable for faceting
}
\note{
Adapted from https://www.mathiasecology.com/code/add-tukeys-significant-letters-to-ggplots
}
\examples{
data <- data.frame("Category" = c(rep("Low", 10), rep("Medium", 10), rep("High", 10)),
                        "Value" = c(rnorm(10, 5), rnorm(10, 5.5), rnorm(10, 10)))
boxplot_letters(data, Category, Value)
}
\author{
Ethan Bass
}
