% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tukey.R
\name{geom_tukey}
\alias{geom_tukey}
\title{Create Compact Letter Display Layer
Groups with at least one letter in common are not significantly different.}
\usage{
geom_tukey(
  test = c("tukey", "kruskalmc", "dunn"),
  type = c("two-way", "one-way"),
  threshold = 0.05,
  where = c("box", "whisker", "mean", "median", "se", "sd", "cl_normal", "cl_boot"),
  hjust = 0,
  vjust = -0.2,
  geom = "text",
  size = 4,
  color = "black",
  fill = "white",
  alpha = 1,
  na.rm = TRUE,
  reversed = FALSE
)
}
\arguments{
\item{test}{Which test to run for pairwise comparisons. Either \code{tukey}
(the default), \code{\link[pgirmess]{kruskalmc}}, or
\code{\link[rstatix]{dunn_test}}.}

\item{type}{If a grouping variable is provided, determines whether to run
separate tests for each facet (\code{one-way}) or one (\code{two-way}) test with
an interaction term between \code{x} and \code{group}. Defaults to \code{two-way}.}

\item{threshold}{Statistical threshold for significance. Defaults to 0.05.}

\item{where}{Where to put the letters. Either above the box (\code{box}) or
upper whisker (\code{whisker}) of a boxplot; at the \code{mean} or
\code{median}; or at the top of the error bars calculated from the standard
error (\code{se}), standard deviation \code{sd}, or 95\% confidence intervals
returned by \code{\link[Hmisc]{smean.cl.normal}}, or \code{\link[Hmisc]{smean.cl.boot}}.}

\item{hjust}{Horizontal adjustment of the label. (Argument to
\code{\link[ggplot2]{geom_text}}).}

\item{vjust}{Vertical adjustment of the label. (Argument to
\code{\link[ggplot2]{geom_text}}).}

\item{geom}{Which geom to use to plot letters. Options are \code{text} and
\code{label}.}

\item{size}{Label size. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{color}{Label color.}

\item{fill}{Label fill (only applies if \code{geom == "label"}).}

\item{alpha}{Label transparency. Defaults to 1.}

\item{na.rm}{Logical. Whether to remove observations with NAs for the provided
factors (i.e. \code{x} and \code{group}) before plotting. Defaults to
\code{TRUE}.}

\item{reversed}{Logical. Argument to
\code{\link[multcompView]{multcompLetters3}}. Determines whether order of
letters should be reversed. Defaults to \code{FALSE}.}
}
\description{
Create Compact Letter Display Layer
Groups with at least one letter in common are not significantly different.
}
\note{
Thank you to \href{https://github.com/yutannihilation}{Hiroaki Yutani}
and \href{https://github.com/simonpcouch}{Simon P. Couch} for a couple of
very helpful blog posts (\href{https://yutani.rbind.io/post/2017-11-07-ggplot-add/}{1},
\href{https://www.simonpcouch.com/blog/ggplot-pipe-plus/}{2}) describing the
\code{\link[ggplot2]{ggplot_add}} syntax.
}
\examples{
library(ggplot2)
set.seed(1)
data <- data.frame("Category" = c(rep("Low", 10), rep("Medium", 10), rep("High", 10)),
                  "Value" = c(rnorm(10, 5), rnorm(10, 5.5), rnorm(10, 10)),
                  "Size" = c("Big","Small"))
data |> ggplot(aes(x=Category, y=Value)) + geom_boxplot() + facet_wrap(~Size) + geom_tukey()
data |> ggplot(aes(x=Size, y=Value)) + geom_boxplot() + facet_wrap(~Category) + geom_tukey()
}
\references{
\itemize{
\item Piepho, Hans-Peter. An Algorithm for a Letter-Based Representation of
All-Pairwise Comparisons. Journal of Computational and Graphical Statistics
13, no. 2 (June 1, 2004): 456–66. \doi{10.1198/1061860043515}.
\item Piepho, Hans-Peter. “Letters in Mean Comparisons: What They Do and Don’t Mean.”
Agronomy Journal 110, no. 2 (2018): 431–34. \doi{10.2134/agronj2017.10.0580}
\item Graves S, Piepho H, Dorai-Raj LSwhfS (2019). multcompView: Visualizations
of Paired Comparisons. R package version 0.1-8. \url{https://CRAN.R-project.org/package=multcompView}
}
}
\author{
Ethan Bass
}
