% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tukey.R
\name{geom_tukey}
\alias{geom_tukey}
\title{Tukey Geom}
\usage{
geom_tukey(
  test = "tukey",
  type = c("global", "local"),
  where = c("box", "whisker", "mean", "median"),
  hjust = 0,
  vjust = 0,
  size = 4,
  na.rm = TRUE
)
}
\arguments{
\item{test}{Which test to run for pairwise comparisons. Default is \code{tukey}.}

\item{type}{If a grouping variable is provided, determines whether to run
separate tests for each facet (\code{local}) or one (\code{global}) test with
an interaction term between \code{x} and \code{group}. Defaults to \code{global}.}

\item{where}{Where to put the letters. Either above the box (\code{box}) or
above the upper whisker (\code{whisker}).}

\item{hjust}{Horizontal adjustment of label. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{vjust}{Vertical adjustment of label. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{size}{Label size. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{na.rm}{Logical. Whether to remove observations with NAs for the provided
factors (i.e. \code{x} and \code{group}) before plotting.}
}
\description{
Tukey Geom
}
\note{
Thank you to Hiroaki Yutani for a very helpful blog post describing the
ggplot_add syntax (https://yutani.rbind.io/post/2017-11-07-ggplot-add/).
}
\examples{
library(ggplot2)
set.seed(1)
data <- data.frame("Category" = c(rep("Low", 10), rep("Medium", 10), rep("High", 10)),
                  "Value" = c(rnorm(10, 5), rnorm(10, 5.5), rnorm(10, 10)),
                  "Size" = c("Big","Small"))
data |> ggplot(aes(x=Category, y=Value)) + geom_boxplot() + facet_wrap(~Size) + geom_tukey()
data |> ggplot(aes(x=Size, y=Value)) + geom_boxplot() + facet_wrap(~Category) + geom_tukey()
}
\author{
Ethan Bass
}
