% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tukey.R
\name{geom_tukey}
\alias{geom_tukey}
\title{Create Compact Letter Display Layer
Groups with at least one letter in common are not significantly different.}
\usage{
geom_tukey(
  test = "tukey",
  type = c("global", "local"),
  where = c("box", "whisker", "mean", "median"),
  hjust = 0,
  vjust = 0,
  size = 4,
  na.rm = TRUE,
  threshold = 0.05
)
}
\arguments{
\item{test}{Which test to run for pairwise comparisons. Default is \code{tukey}.}

\item{type}{If a grouping variable is provided, determines whether to run
separate tests for each facet (\code{local}) or one (\code{global}) test with
an interaction term between \code{x} and \code{group}. Defaults to \code{global}.}

\item{where}{Where to put the letters. Either above the box (\code{box}) or
above the upper whisker (\code{whisker}).}

\item{hjust}{Horizontal adjustment of label. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{vjust}{Vertical adjustment of label. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{size}{Label size. Argument to \code{\link[ggplot2]{geom_text}}.}

\item{na.rm}{Logical. Whether to remove observations with NAs for the provided
factors (i.e. \code{x} and \code{group}) before plotting.}

\item{threshold}{Statistical threshold for significance. Defaults to 0.05.}
}
\description{
Create Compact Letter Display Layer
Groups with at least one letter in common are not significantly different.
}
\note{
Thank you to Hiroaki Yutani ((https://yutani.rbind.io/post/2017-11-07-ggplot-add/))
and Simon P Couch (https://www.simonpcouch.com/blog/ggplot-pipe-plus/) for two
very helpful blog posts describing the \code{\link[ggplot2]{ggplot_add}} syntax.
}
\examples{
library(ggplot2)
set.seed(1)
data <- data.frame("Category" = c(rep("Low", 10), rep("Medium", 10), rep("High", 10)),
                  "Value" = c(rnorm(10, 5), rnorm(10, 5.5), rnorm(10, 10)),
                  "Size" = c("Big","Small"))
data |> ggplot(aes(x=Category, y=Value)) + geom_boxplot() + facet_wrap(~Size) + geom_tukey()
data |> ggplot(aes(x=Size, y=Value)) + geom_boxplot() + facet_wrap(~Category) + geom_tukey()
}
\references{
\itemize{
\item Piepho, Hans-Peter (2004) "An Algorithm for a Letter-Based
Representation of All-Pairwise Comparisons", Journal of Computational and
Graphical Statistics, 13(2)456-466.
\item Graves S, Piepho H, Dorai-Raj LSwhfS (2019). \emph{multcompView: Visualizations of Paired Comparisons}. R package
version 0.1-8, (https://CRAN.R-project.org/package=multcompView).
\item Piepho, Hans-Peter. “Letters in Mean Comparisons: What They Do and Don’t Mean.”
Agronomy Journal 110, no. 2 (2018): 431–34. \doi{10.2134/agronj2017.10.0580}/
}
}
\author{
Ethan Bass
}
